engine: delete all remaining PhysX — engine is now AZSL/host-only
The PhysX→AZSL swap left the production path clean but the legacy PhysX
scaffolding standing. None of it did production work; remove it all:

- Delete the dead bond pipeline: HCPParticlePipeline::Disassemble/Reassemble
  (zero callers). Gut HCPParticlePipeline.h to the Bond/PBMData data types
  (still used by PbmWriter/Reader/BondCompiler); delete HCPParticlePipeline.cpp.
- Delete the legacy PhysX byte→char/word trials: HCPSuperpositionTrial.{h,cpp}
  and HCPWordSuperpositionTrial.{h,cpp}. The shared CharRun/RunTag types they
  defined move to a new HCPCharRun.h (repointed: ResolutionChamber, ByteIngest).
- Remove their only entry points: the 3 SourcePhys* console commands and the
  already-deprecated socket phys_resolve endpoint (its body was dead after an
  early return).
- Drop the vestigial pipeline IsInitialized() gate in HCPPhysIngest.
- Remove the PhysX/CUDA boot init from HCPEngineSystemComponent::Activate()
  (the m_particlePipeline member, accessor, Shutdown, IsReady/IsEngineReady now
  key off vocab + BedManager). No more PhysX system acquisition at boot.
- Remove the Gem::PhysX5.Static build dependency and the PhysX5 gem.json
  dependency.

Fixed latent missing-includes the unity-build regrouping exposed (self-
sufficient headers): lmdb.h + unordered_set in HCPVocabBed.h, unordered_set in
HCPEnvelopeManager.cpp.

libHCPEngine.so links clean with PhysX removed; nm shows zero PhysX symbols.
Settle ctest green.
