+++ hcp-engine/Gem/CMakeLists.txt
@@ -54,7 +54,6 @@ ly_add_target(
         PRIVATE
             AZ::AzGameFramework
             Gem::Atom_AtomBridge.Static
-            Gem::PhysX5.Static
             ${LIBPQ_LIBRARY}
             ${LIBLMDB_LIBRARY}
             ${LIBSQLITE3_LIBRARY}
