+++ hcp-engine/Gem/Source/HCPParticlePipeline.h
@@ -1,24 +1,15 @@
 #pragma once
 
 #include <AzCore/std/containers/vector.h>
-#include <AzCore/std/containers/unordered_map.h>
 #include <AzCore/std/string/string.h>
 
-// Forward declarations — full PhysX headers only in .cpp
-namespace physx
-{
-    class PxPhysics;
-    class PxFoundation;
-    class PxScene;
-    class PxPBDParticleSystem;
-    class PxParticleMaterial;
-    class PxCudaContextManager;
-}
+// Pair-bond data types. These were once produced by a PhysX PBD "particle
+// pipeline" (disassembly: text -> bonds); that PhysX path has been removed in
+// the AZSL swap. The types remain as the stored form of a document, written and
+// read by HCPPbmWriter / HCPPbmReader and produced by the manifest scanner.
 
 namespace HCPEngine
 {
-    class HCPVocabulary;
-
     //! A bond triple: (token_a, token_b, count)
     struct Bond
     {
@@ -29,8 +20,7 @@ namespace HCPEngine
 
     //! Pair Bond Map — the stored form of a document.
     //! Each bond is (A, B, count) representing adjacent token pair occurrences.
-    //! The set of bonds IS the document. Physics reassembly recovers the
-    //! unique ordering from the bond constraints.
+    //! The set of bonds IS the document.
     struct PBMData
     {
         AZStd::vector<Bond> bonds;
@@ -38,59 +28,6 @@ namespace HCPEngine
         AZStd::string firstFpbB;  // First forward pair bond B-side
         size_t totalPairs = 0;
         size_t uniqueTokens = 0;
-    };
-
-    //! The particle pipeline: manages PhysX PBD particle system for
-    //! disassembly (text -> bonds) and reassembly (bonds -> text).
-    class HCPParticlePipeline
-    {
-    public:
-        HCPParticlePipeline() = default;
-        ~HCPParticlePipeline();
-
-        //! Initialize PhysX PBD subsystem: CUDA context, GPU scene, particle system.
-        bool Initialize(physx::PxPhysics* pxPhysics, physx::PxFoundation* pxFoundation);
-
-        //! Shut down PBD subsystem and release resources
-        void Shutdown();
-
-        //! Run disassembly: token sequence -> pair bond map via PBD particle physics.
-        PBMData Disassemble(const AZStd::vector<AZStd::string>& tokenIds);
-
-        //! Run reassembly: bond map -> ordered token sequence via PBD particle physics.
-        //! Dumbbells dissolve when both ends find their mates.
-        //! Whitespace is re-inserted at dissolution points unless
-        //! stickiness rules suppress it (punctuation sticks to adjacent tokens).
-        //! @return Ordered string sequence including whitespace
-        AZStd::vector<AZStd::string> Reassemble(
-            const PBMData& pbmData,
-            const HCPVocabulary& vocab);
-
-        bool IsInitialized() const { return m_initialized; }
-
-        //! Create a dedicated GPU-enabled PxScene for char→word resolution.
-        //! Separate from the main scene (byte→char / disassembly).
-        //! One simulate() on this scene processes all resolution chambers.
-        bool CreateCharWordScene();
-
-        // Resource accessors — for detection scene and other physics operations
-        physx::PxPhysics* GetPhysics() const { return m_pxPhysics; }
-        physx::PxScene* GetScene() const { return m_pxScene; }
-        physx::PxScene* GetCharWordScene() const { return m_charWordScene; }
-        physx::PxCudaContextManager* GetCuda() const { return m_cudaContextManager; }
-
-    private:
-        bool m_initialized = false;
-
-        physx::PxPhysics* m_pxPhysics = nullptr;
-        physx::PxCudaContextManager* m_cudaContextManager = nullptr;
-        physx::PxScene* m_pxScene = nullptr;
-        physx::PxScene* m_charWordScene = nullptr;  // Dedicated Phase 2 scene
-        physx::PxParticleMaterial* m_particleMaterial = nullptr;
-        physx::PxParticleMaterial* m_reassemblyMaterial = nullptr;
-
-        // Working data for disassembly bond counting
-        AZStd::unordered_map<AZStd::string, int> m_bondCounts;
     };
 
 } // namespace HCPEngine
