+++ hcp-engine/Gem/Source/HCPSocketServer.cpp
@@ -5,8 +5,6 @@
 #include "HCPParticlePipeline.h"
 #include "HCPResolutionChamber.h"
 #include "HCPVocabBed.h"
-#include "HCPSuperpositionTrial.h"
-#include "HCPWordSuperpositionTrial.h"
 #include "HCPBondCompiler.h"
 #include "HCPEnvelopeManager.h"
 #include "HCPPhysIngest.h"
@@ -586,188 +584,10 @@ namespace HCPEngine
             return AZStd::string(sb.GetString(), sb.GetSize());
         }
 
-        // ---- phys_resolve — DEPRECATED: use phys_ingest ----
+        // ---- phys_resolve — REMOVED: PhysX path deleted; use phys_ingest ----
         if (strcmp(action, "phys_resolve") == 0)
         {
             return R"({"status":"error","message":"phys_resolve is deprecated — use phys_ingest"})";
-            // Dead code below retained for reference only
-            AZStd::string text;
-            AZ::u32 maxChars = 0;  // 0 = process entire text
-
-            if (doc.HasMember("file") && doc["file"].IsString())
-            {
-                AZStd::string filePath(doc["file"].GetString(), doc["file"].GetStringLength());
-                std::ifstream ifs(filePath.c_str());
-                if (!ifs.is_open())
-                {
-                    return R"({"status":"error","message":"Could not open file"})";
-                }
-                std::string stdText((std::istreambuf_iterator<char>(ifs)),
-                                     std::istreambuf_iterator<char>());
-                ifs.close();
-                text = AZStd::string(stdText.c_str(), stdText.size());
-            }
-            else if (doc.HasMember("text") && doc["text"].IsString())
-            {
-                text = AZStd::string(doc["text"].GetString(), doc["text"].GetStringLength());
-            }
-            else
-            {
-                return R"({"status":"error","message":"phys_resolve requires 'file' or 'text'"})";
-            }
-
-            if (doc.HasMember("max_chars") && doc["max_chars"].IsUint())
-            {
-                maxChars = doc["max_chars"].GetUint();
-            }
-
-            HCPParticlePipeline& pipeline = m_engine->GetParticlePipeline();
-            if (!pipeline.IsInitialized())
-            {
-                return R"({"status":"error","message":"Particle pipeline not initialized"})";
-            }
-
-            // Phase 1: byte→char settlement
-            SuperpositionTrialResult phase1 = RunSuperpositionTrial(
-                pipeline.GetPhysics(),
-                pipeline.GetScene(),
-                pipeline.GetCuda(),
-                text, m_engine->GetVocabulary(), maxChars);
-
-            fprintf(stderr, "[phys_resolve] Phase 1: %u/%u settled (%.1f%%) in %.1f ms [%u bytes → %u codepoints]\n",
-                phase1.settledCount, phase1.totalCodepoints,
-                phase1.totalCodepoints > 0 ? 100.0f * phase1.settledCount / phase1.totalCodepoints : 0.0f,
-                phase1.simulationTimeMs, phase1.totalBytes, phase1.totalCodepoints);
-            fflush(stderr);
-
-            // Extract character runs from Phase 1 output
-            AZStd::vector<CharRun> runs = ExtractRunsFromCollapses(phase1);
-            if (runs.empty())
-            {
-                return R"({"status":"error","message":"No runs extracted from Phase 1 output"})";
-            }
-
-            fprintf(stderr, "[phys_resolve] Extracted %zu runs from Phase 1 output (max %u bytes)\n",
-                runs.size(), maxChars);
-            fflush(stderr);
-
-            // Use persistent BedManager (initialized at Activate)
-            BedManager& bedManager = m_engine->GetBedManager();
-            if (!bedManager.IsInitialized())
-            {
-                return R"({"status":"error","message":"BedManager not initialized"})";
-            }
-
-            ResolutionManifest manifest = bedManager.Resolve(runs);
-
-            // ---- Optional benchmark TSV output ----
-            bool writeBenchmark = doc.HasMember("benchmark") &&
-                doc["benchmark"].IsBool() && doc["benchmark"].GetBool();
-            if (writeBenchmark)
-            {
-                struct rusage usage;
-                getrusage(RUSAGE_SELF, &usage);
-                long rssKb = usage.ru_maxrss;
-
-                long vramUsedMb = 0;
-                FILE* nvsmi = popen("nvidia-smi --query-gpu=memory.used --format=csv,noheader,nounits 2>/dev/null", "r");
-                if (nvsmi)
-                {
-                    char buf[64];
-                    if (fgets(buf, sizeof(buf), nvsmi))
-                        vramUsedMb = atol(buf);
-                    pclose(nvsmi);
-                }
-
-                mkdir("benchmarks", 0755);
-
-                auto now = std::chrono::system_clock::now();
-                auto epochMs = std::chrono::duration_cast<std::chrono::milliseconds>(
-                    now.time_since_epoch()).count();
-
-                char tsvPath[256];
-                snprintf(tsvPath, sizeof(tsvPath), "benchmarks/resolve_%lld.tsv", (long long)epochMs);
-
-                FILE* tsv = fopen(tsvPath, "w");
-                if (tsv)
-                {
-                    fprintf(tsv, "timestamp\tbytes\tphase1_codepoints\tphase1_settled\t"
-                        "phase1_pct\tphase1_ms\ttotal_runs\tresolved\tunresolved\t"
-                        "resolved_pct\tresolve_ms\trss_kb\tvram_mb\n");
-                    fprintf(tsv, "%lld\t%zu\t%u\t%u\t%.1f\t%.1f\t%u\t%u\t%u\t%.1f\t%.1f\t%ld\t%ld\n",
-                        (long long)epochMs,
-                        text.size(),
-                        phase1.totalCodepoints,
-                        phase1.settledCount,
-                        phase1.totalCodepoints > 0 ? 100.0f * phase1.settledCount / phase1.totalCodepoints : 0.0f,
-                        static_cast<double>(phase1.simulationTimeMs),
-                        manifest.totalRuns,
-                        manifest.resolvedRuns,
-                        manifest.unresolvedRuns,
-                        manifest.totalRuns > 0 ? 100.0f * manifest.resolvedRuns / manifest.totalRuns : 0.0f,
-                        static_cast<double>(manifest.totalTimeMs),
-                        rssKb,
-                        vramUsedMb);
-                    fclose(tsv);
-                    fprintf(stderr, "[phys_resolve] Benchmark written: %s\n", tsvPath);
-                    fflush(stderr);
-                }
-            }
-
-            // Build JSON response
-            rapidjson::StringBuffer sb;
-            rapidjson::Writer<rapidjson::StringBuffer> w(sb);
-            w.StartObject();
-            w.Key("status"); w.String("ok");
-            w.Key("phase1_settled"); w.Uint(phase1.settledCount);
-            w.Key("phase1_total"); w.Uint(phase1.totalCodepoints);
-            w.Key("phase1_total_bytes"); w.Uint(phase1.totalBytes);
-            w.Key("phase1_time_ms"); w.Double(static_cast<double>(phase1.simulationTimeMs));
-            w.Key("total_runs"); w.Uint(manifest.totalRuns);
-            w.Key("resolved"); w.Uint(manifest.resolvedRuns);
-            w.Key("unresolved"); w.Uint(manifest.unresolvedRuns);
-            w.Key("time_ms"); w.Double(static_cast<double>(manifest.totalTimeMs));
-            w.Key("bed_initialized"); w.Bool(m_engine->GetBedManager().IsInitialized());
-
-            // System resource snapshot
-            {
-                struct rusage usage;
-                getrusage(RUSAGE_SELF, &usage);
-                w.Key("rss_kb"); w.Int64(usage.ru_maxrss);
-            }
-
-            w.Key("results");
-            w.StartArray();
-            for (const auto& r : manifest.results)
-            {
-                w.StartObject();
-                w.Key("run"); w.String(r.runText.c_str());
-                w.Key("resolved"); w.Bool(r.resolved);
-                if (r.resolved)
-                {
-                    w.Key("word"); w.String(r.matchedWord.c_str());
-                    w.Key("token_id"); w.String(r.matchedTokenId.c_str());
-                    w.Key("tier"); w.Uint(r.tierResolved);
-                    if (r.morphBits != 0)
-                        { w.Key("morph_bits"); w.Uint(r.morphBits); }
-                }
-                if (r.firstCap)
-                    { w.Key("first_cap"); w.Bool(true); }
-                if (r.allCaps)
-                    { w.Key("all_caps"); w.Bool(true); }
-                w.EndObject();
-            }
-            w.EndArray();
-
-            w.EndObject();
-
-            fprintf(stderr, "[phys_resolve] Complete: %u/%u resolved (%.1f%%) in %.1f ms\n",
-                manifest.resolvedRuns, manifest.totalRuns,
-                manifest.totalRuns > 0 ? 100.0f * manifest.resolvedRuns / manifest.totalRuns : 0.0f,
-                manifest.totalTimeMs);
-            fflush(stderr);
-
-            return AZStd::string(sb.GetString(), sb.GetSize());
         }
 
         // ---- phys_ingest (Phase 1 + Phase 2 + scanner → PBM to Postgres) ----
